stackmachine: Fix lookahead on short reads
This was a schrödingbug:
Not encountered until seen in the code.

The expression for is_horizon_lengthenable
was slightly wrong: It would only be wrong
on a short read (shorter than requested).
Short reads practically don't happen
unless reading from a slow pipe. But the bug
was indeed possible to awaken by doing so.

Background:

Bourne shell has ambiguous grammar:
A beginning of a word in command position
could be the beginning of an assignment,
keyword, or, by nullhypothesis, a command.
Thus why lookahead is necessary.

Shellharden does not read the whole file
into memory at once, but works more like
a filter: It can work in an endless pipe.
It has a lookahead buffer – horizon – that
is moved forward lazily. The horizon can
only be moved forward if not EOF or full.

Bug & impact:

The problem is not that the horizon is
short, but in the permission to extend it:
Parse contexts could effectively be told
that this is it, as if EOF in the middle
of a file. This forces parse contexts to
reject ideas about things coming after.
This affects the effective parse tree,
but only on a high level that is not
catastrophic to what Shellharden does:
There is at worst the nullhypotheis
that everything is a command,
and everything must be quoted.