(feat) dockerize shellharden
Enables a user to run shellharden in a Docker
container.

- Add Dockerfile & entrypoint
= Update README.md
    - Add Docker usage
    - Wrap lines to make it easier to read from CLI
    - Correct misc. grammar & punctuation

I did this so I can use it in a CI/CD pipeline, but thought
it might be worth contributing back. If you want, I can publish
to dockerhub before you accept this PR.

Thanks for such a great tool!

Signed-off-by: Vladislav Doster <mvdoster@gmail.com>