treatfile: Put fallible section in a function
Refactoring for the next commit. No functional change.

Fallible section is what I call the technique of wrapping a section of code
in a control structure for the ability to transfer control back out to the
same point from multiple points within, as would be achieved with gotos
that skip ahead to the same label, typically for resource cleanup.
Using destructors is only logically equivalent, as this instead duplicates
the cleanup section (of implicit destructor calls) at each return point.
Inlining the fallible section achieves the desired control flow.