Fix misdetection of line continuation as a command also in case arms
Addendum to commit 7454c3412:
Moves the fix to a helper function used both places
(see the expanded test case).

SitCaseArm is a special case of SitNormal
to handle the syntactic peculiarities of case arms.
The challenge with special cases is to avoid
the hazard of needing to fix everything twice.
Not because of duplication (I use helper functions),
but having the fantasy to look for another case:
You don't discover that both need the same fix
at the same time, so fixing it in shared code
is only right when it becomes so.